refactor!: move exceptions to their own folder
BREAKING CHANGE: All exceptions are moved to under the new `Eflatun.SceneReference.Exceptions` namespace. Add using directives to your code in order to upgrade.
