+++ Eflatun.SceneReference/Assets/Demo/Validation/ValidationDemo.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Text;
+using Eflatun.SceneReference.Exceptions;
 using UnityEngine;
 using UnityEngine.Assertions;
