+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/AddressableSceneGuidToAddressMapProvider.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using System.Linq;
+using Eflatun.SceneReference.Exceptions;
 using JetBrains.Annotations;
 using UnityEngine;
 using UnityEngine.Scripting;
