+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 5daedf7ef42e488b88d8bc66745083a4
+timeCreated: 1686399518
