+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions/AddressNotFoundException.cs
@@ -2,7 +2,7 @@
 using System.Runtime.Serialization;
 using JetBrains.Annotations;
 
-namespace Eflatun.SceneReference
+namespace Eflatun.SceneReference.Exceptions
 {
     /// <summary>
     /// Thrown if an address is not found in the addressable scene guid to address map.
