+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions/AddressNotUniqueException.cs
@@ -2,7 +2,7 @@
 using System.Runtime.Serialization;
 using JetBrains.Annotations;
 
-namespace Eflatun.SceneReference
+namespace Eflatun.SceneReference.Exceptions
 {
     /// <summary>
     /// Thrown if an address matches multiple assets.
