+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions/SceneReferenceException.cs
@@ -2,7 +2,7 @@
 using System.Runtime.Serialization;
 using JetBrains.Annotations;
 
-namespace Eflatun.SceneReference
+namespace Eflatun.SceneReference.Exceptions
 {
     /// <summary>
     /// Root custom exception for the <c>Eflatun.SceneReference</c> package.
