+++ Eflatun.SceneReference/Packages/com.eflatun.scenereference/Runtime/Exceptions/SceneReferenceInternalException.cs.meta
